+++ src/main/java/de/codescape/jira/plugins/scrumpoker/rest/mapper/SessionEntityMapper.java
@@ -14,7 +14,6 @@ import de.codescape.jira.plugins.scrumpoker.model.AllowRevealDeck;
 import de.codescape.jira.plugins.scrumpoker.rest.entities.*;
 import de.codescape.jira.plugins.scrumpoker.service.CardSetService;
 import de.codescape.jira.plugins.scrumpoker.service.GlobalSettingsService;
-import org.apache.commons.lang3.math.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -215,7 +214,7 @@ public class SessionEntityMapper {
             return false;
         if (scrumPokerSession.getVotes().length == 1)
             return false;
-        if (!isAssignableToEstimationField(vote.getVote()))
+        if (!isAssignableToEstimateField(vote.getVote()))
             return true;
         Integer current = Integer.valueOf(vote.getVote());
         return current.equals(getMaximumVote(scrumPokerSession.getVotes()))
@@ -264,7 +263,7 @@ public class SessionEntityMapper {
      */
     private BoundedVoteEntity createBoundedVote(String value, Map<String, Long> distribution) {
         return new BoundedVoteEntity(value, distribution.getOrDefault(value, 0L),
-            isAssignableToEstimationField(value));
+            isAssignableToEstimateField(value));
     }
 
     /**
@@ -287,7 +286,7 @@ public class SessionEntityMapper {
     @Deprecated
     private List<Integer> numericValues(ScrumPokerVote[] votes) {
         return stream(votes).map(ScrumPokerVote::getVote)
-            .filter(this::isAssignableToEstimationField)
+            .filter(this::isAssignableToEstimateField)
             .map(Integer::valueOf)
             .collect(Collectors.toList());
     }
@@ -306,8 +305,8 @@ public class SessionEntityMapper {
     /**
      * Returns whether the given vote can be assigned to the estimation field.
      */
-    private boolean isAssignableToEstimationField(String vote) {
-        return NumberUtils.isNumber(vote);
+    private boolean isAssignableToEstimateField(String vote) {
+        return !isSpecialCardValue(vote);
     }
 
 }